+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Improved
+- added courier shipping method for Cyprus and Malta
+
 ## [1.7.1] - 2025-12-02
 ### Fixed
 - fixed label download issue that occurred after update 1.7.0